SimpleDrawing_FlyAI/requirements.txt
flyai
numpy
Pillow
torch==1.2.0
torchvision==0.4.0
timm
